cf_load_to_bq/requirements.txt
google-cloud-storage==1.12.0
google-cloud-bigquery==1.5.1
